python/pyproject.toml
[project]
name = "azure-function-durable-python"
version = "0.1.0"
description = "Azure Durable Functions implementation in Python"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "azure-functions>=1.18.0",
    "azure-functions-durable>=1.2.0",
]

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
